fix(theme): 落实 CodeRabbit 评审——注册清理兜底、快照先过滤后排序、共享运行时快照 hook
- tuiThemes.register：bindCallerEffect 失败时先 dispose 再返回 NOOP，
  避免插件卸载后主题残留可解析
- themeCatalog：先过滤非字符串 name 再排序（比较器不再能抛 TypeError），
  排序键改用 keyOf 小写，摆脱运行时 ICU locale 依赖
- ThemeProvider/ThemePicker 的运行时主题 store 适配器抽为共享
  useRuntimeThemeSnapshot hook，两处回退行为不再可能漂移
- /theme 直切去掉恒等于 parts[0] 的目录查找，setTheme 自身已校验
- verify-runtime-themes：稳定身份断言存变量比较，避开 noSelfCompare
- theme-i18n 审计文档标注 CC_TUI_*/~/.dsh-cc 为基线时期旧名并指向现名
